requirements.txt
alabaster==0.7.12
appdirs==1.4.3
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
black==18.9b0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
docutils==0.14
gitpython==2.1.11
idna==2.7
imagesize==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
latexcodec==1.0.5
MarkupSafe==1.0
matplotlib==3.0.0
more-itertools==4.3.0
networkx==2.3
numpy==1.15.2
oset==0.1.3
packaging==18.0
pandas==0.24.2
pluggy==0.7.1
py==1.7.0
pyang==1.7.8
pyangbind==0.8.1
pybtex==0.21
pybtex-docutils==0.2.1
Pygments==2.2.0
pyparsing==2.2.2
pytest==3.8.2
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
scipy==1.1.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-bibtex==0.4.0
sphinxcontrib-websupport==1.1.0
toml==0.10.0
urllib3==1.23
xlrd==1.2.0
